Move specifications plugin to builtin_commands package
